# ui/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Quantix",
    "version": "0.1.0"
  },
  "paths": {
...
    "/tenders/{tender_id}/documents": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Add Documents",
        "operationId": "add_documents_tenders__tender_id__documents_post",
        "parameters": [
          {
            "name": "tender_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tender Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_add_documents_tenders__tender_id__documents_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Added"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "List Documents",
        "operationId": "list_documents_tenders__tender_id__documents_get",
        "parameters": [
          {
            "name": "tender_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tender Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentOut"
                  },
                  "title": "Response List Documents Tenders  Tender Id  Documents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tenders/{tender_id}/documents/import": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Import Documents",
        "description": "Copy the chosen files into the tender from where they are, as the desktop app's own pickers give them. The\nengineer's files are only read, never changed.",
        "operationId": "import_documents_tenders__tender_id__documents_import_post",
        "parameters": [
          {
            "name": "tender_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tender Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Added"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tenders/{tender_id}/search": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Search",
        "operationId": "search_tenders__tender_id__search_get",
        "parameters": [
          {
            "name": "tender_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tender Id"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Q"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchHit"
                  },
                  "title": "Response Search Tenders  Tender Id  Search Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/documents/{document_id}/pages/{number}": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Get Page",
        "operationId": "get_page_documents__document_id__pages__number__get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          },
          {
            "name": "number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Number"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/quantix__api__documents__PageOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/documents/{document_id}/pages/{number}/image": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Page Image",
        "operationId": "page_image_documents__document_id__pages__number__image_get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          },
          {
            "name": "number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Number"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/documents/{document_id}/file": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Original File",
        "operationId": "original_file_documents__document_id__file_get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/documents/{document_id}/sheets/{number}": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Sheet",
        "description": "A sheet of a workbook as Excel shows it: its cells' text, merges, sizes and formatting. With `hidden`, the rows\nand columns the sheet hides are shown too.",
        "operationId": "sheet_documents__document_id__sheets__number__get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          },
          {
            "name": "number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Number"
            }
          },
          {
            "name": "hidden",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Hidden"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SheetView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/documents/{document_id}/open": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Open In App",
        "description": "Opens the file in the app this computer uses for its type. The app gets a read-only copy, so the stored file\nstays exactly as it was supplied.",
        "operationId": "open_in_app_documents__document_id__open_post",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Added": {
        "properties": {
          "added": {
            "type": "integer",
            "title": "Added"
          },
          "unchanged": {
            "type": "integer",
            "title": "Unchanged"
          }
        },
        "type": "object",
        "required": [
          "added",
          "unchanged"
        ],
        "title": "Added"
      },
...
      "Body_add_documents_tenders__tender_id__documents_post": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Files"
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "Body_add_documents_tenders__tender_id__documents_post"
      },
...
      "Cell": {
        "properties": {
          "column": {
            "type": "integer",
            "title": "Column"
          },
          "text": {
            "type": "string",
            "title": "Text",
            "default": ""
          },
          "style": {
            "type": "integer",
            "title": "Style",
            "default": 0
          },
          "rows": {
            "type": "integer",
            "title": "Rows",
            "default": 1
          },
          "columns": {
            "type": "integer",
            "title": "Columns",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "column"
        ],
        "title": "Cell"
      },
...
      "Column": {
        "properties": {
          "letter": {
            "type": "string",
            "title": "Letter"
          },
          "width": {
            "type": "number",
            "title": "Width"
          },
          "hidden": {
            "type": "boolean",
            "title": "Hidden",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "letter",
          "width"
        ],
        "title": "Column"
      },
...
      "DocumentOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "page_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Count"
          },
          "group_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "scans_to_read": {
            "type": "integer",
            "title": "Scans To Read",
            "default": 0
          },
          "opened": {
            "type": "integer",
            "title": "Opened",
            "default": 0
          },
          "cited": {
            "type": "integer",
            "title": "Cited",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "path",
          "name",
          "kind",
          "size",
          "status",
          "note",
          "page_count",
          "group_name",
          "description"
        ],
        "title": "DocumentOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImportIn": {
        "properties": {
          "folder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Folder"
          },
          "files": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Files",
            "default": []
          }
        },
        "type": "object",
        "title": "ImportIn",
        "description": "What the engineer chose on this computer: a whole folder, kept with its folders, or single files."
      },
...
      "Row": {
        "properties": {
          "number": {
            "type": "integer",
            "title": "Number"
          },
          "height": {
            "type": "number",
            "title": "Height"
          },
          "hidden": {
            "type": "boolean",
            "title": "Hidden",
            "default": false
          },
          "cells": {
            "items": {
              "$ref": "#/components/schemas/Cell"
            },
            "type": "array",
            "title": "Cells"
          }
        },
        "type": "object",
        "required": [
          "number",
          "height",
          "cells"
        ],
        "title": "Row"
      },
...
      "SearchHit": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "snippet": {
            "type": "string",
            "title": "Snippet"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "name",
          "page",
          "snippet"
        ],
        "title": "SearchHit"
      },
...
      "SheetTab": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "hidden": {
            "type": "boolean",
            "title": "Hidden",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "SheetTab"
      },
      "SheetView": {
        "properties": {
          "sheets": {
            "items": {
              "$ref": "#/components/schemas/SheetTab"
            },
            "type": "array",
            "title": "Sheets"
          },
          "number": {
            "type": "integer",
            "title": "Number"
          },
          "right_to_left": {
            "type": "boolean",
            "title": "Right To Left",
            "default": false
          },
          "gridlines": {
            "type": "boolean",
            "title": "Gridlines",
            "default": true
          },
          "frozen_rows": {
            "type": "integer",
            "title": "Frozen Rows",
            "default": 0
          },
          "frozen_columns": {
            "type": "integer",
            "title": "Frozen Columns",
            "default": 0
          },
          "columns": {
            "items": {
              "$ref": "#/components/schemas/Column"
            },
            "type": "array",
            "title": "Columns"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/Row"
            },
            "type": "array",
            "title": "Rows"
          },
          "styles": {
            "items": {
              "$ref": "#/components/schemas/Style"
            },
            "type": "array",
            "title": "Styles"
          },
          "more_rows": {
            "type": "integer",
            "title": "More Rows",
            "default": 0
          },
          "more_columns": {
            "type": "integer",
            "title": "More Columns",
            "default": 0
          },
          "hidden_rows": {
            "type": "integer",
            "title": "Hidden Rows",
            "default": 0
          },
          "hidden_columns": {
            "type": "integer",
            "title": "Hidden Columns",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "sheets",
          "number",
          "columns",
          "rows",
          "styles"
        ],
        "title": "SheetView"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "quantix__api__documents__PageOut": {
        "properties": {
          "number": {
            "type": "integer",
            "title": "Number"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "has_text": {
            "type": "boolean",
            "title": "Has Text"
          }
        },
        "type": "object",
        "required": [
          "number",
          "text",
          "has_text"
        ],
        "title": "PageOut"
      },
...
    }
  }
}